Extract rebuild machinery into client/rebuild.py
